+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Nylas Java SDK Changelog
 
+## [Unreleased]
+
+### Added
+* Added support for `name` attribute in Scheduler Configuration class to allow custom naming of Scheduling Pages
+
 ### [2.8.0] - Release 2025-04-30
 * Added support for Notetaker APIs
 * Added support for Notetaker via the calendar and event APIs